ecommerce/requirements.txt
# These requirements are only necessary when developing on Oscar.

# development
Werkzeug>=1.0,<2.1
django-debug-toolbar>=2.2,<3.6
django-extensions>=2.2,<3.3
psycopg2-binary>=2.8,<2.10

# Sandbox
Whoosh>=2.7,<2.8
django-environ>=0.4,<0.10
django-redis>=4.12,<5.3
pysolr>=3.9,<3.10
redis>=3.5,<4.4
requests>=2.25,<3
# uWSGI>=2.0.19,<2.1
whitenoise>=5.2,<6.3

# Linting
flake8==4.0.1
flake8-debugger==4.1.2
isort==5.10.1

# Helpers
pyprof2calltree>=1.4,<1.5
ipdb>=0.13,<0.14
ipython>=7.12,<9

# Country data
pycountry


django>=3.1,<3.3
# PIL is required for image fields, Pillow is the "friendly" PIL fork
pillow>=6.0
# We use the ModelFormSetView from django-extra-views for the basket page
django-extra-views>=0.13,<0.15
# Search support
django-haystack>=3.0b1
# Treebeard is used for categories
django-treebeard>=4.3.0
# Babel is used for currency formatting
Babel>=1.0,<3.0
# For manipulating search URLs
purl>=0.7
# For phone number field
phonenumbers
django-phonenumber-field>=4.0.0,<7.0.0
# Used for oscar.test.factories
factory-boy>=3.0,<3.3
# Used for automatically building larger HTML tables
django-tables2>=2.3,<2.4
# Used for manipulating form field attributes in templates (eg: add
# a css class)
django-widget-tweaks>=1.4.1

sorl-thumbnail>=12.6,<12.7
easy-thumbnails>=2.7,<2.8
Sphinx>=4.2,<4.3
sphinxcontrib-napoleon==0.7
sphinxcontrib-spelling==7.2.1
sphinx_rtd_theme==1.0.0
sphinx-issues==1.2.0
WebTest>=2.0,<2.1
coverage>=5.4,<5.5
django-webtest>=1.9,<1.10
pytest-django>=3.7,<3.9
pytest-xdist>=2.2,<3
tox>=3.21,<4
freezegun>=1.1,<2

#
django-rosetta
goslate
gunicorn

djangorestframework==3.14.0